Finalized layer for loading recipes (#20)
* Finalized layer for loading recipes

* finished tests

* finished tests

* make changes for PR

+++ app/src/main/java/io/github/mcasper3/prep/data/database/recipe/RecipeDao.kt
@@ -16,7 +16,7 @@ interface RecipeDao {
 
     @Transaction
     @Query("SELECT * FROM Recipes WHERE _id = :id")
-    fun getById(id: Int): Flowable<RecipeWithSteps>
+    fun getById(id: Long): Flowable<RecipeWithSteps>
 
     @Insert
     fun insertAll(vararg recipes: Recipe)
